[REF] Odoo 8.0-10.0: do not require linking environment variables
The entry points of the Odoo docker images require the presence of the
following environment variables:
    - DB_PORT_5432_TCP_ADDR
    - DB_PORT_5432_TCP_PORT
    - DB_ENV_POSTGRES_USER
    - DB_ENV_POSTGRES_PASSWORD

We use them to set the PGHOST, PGPORT, PGUSER and PGPASSWORD environment
variable that will then be read by Odoo in order to connect to the
database server.

As the first ones are automatically created by Docker when we use the linking
system, it's not possible to run Odoo without the linking system and,
for example, with Docker compose.

In order to make it work with Docker compose and keep the backward
compatibility with the linking system, we have to provide sane
fallbacks to the assignation of the PGHOST, PGPORT, PGUSER and
PGPASSWORD environment variable.

We do that in this commit by defaulting the assignation of the postgres
host to "db" and the port to "5432" because, if you follow our image
documentation, that's what you get. We also introduce the handling of
two others environment variable in the Odoo entry point: POSTGRES_USER
and POSTGRES_PASSWORD. They will be used as fallbacks for
DB_ENV_POSTGRES_USER and DB_ENV_POSTGRES_PASSWORD.