[FIX] Odoo 10.0-12.0: add missing packages
In Odoo, ldap, qrcode and vobject python modules are set as required in
odoo/setup.py. On the other hand, they are only recommendend in the
Debian package.

As a consequence, it was difficult to run some Odoo modules that
required one of them.

The Odoo Debian package was adapted in master where they were added as
dependencies in odoo/odoo@2ff49c5fcb584c8215ecbcb0034690e53732fb49 .

With this commit, they are installed in the Docker images.

This is a part of the fix of odoo/odoo#27752.