[IMP] 11.0-13.0: Wait for PostgreSQL before starting Odoo
In some situations, like when a docker compose file use a data volume
to persist postgresql databases, the Odoo server starts and tries to
connect to database server before the postgresql server is ready.

In such a case, the Odoo containers fails to start.

With this commit, the wait-for-it script is used to wait for the
database server service to be ready before starting Odoo.

This script is available as a Debian package.

Closes #258, closes #259